+++ src/main/java/pinup/backend/report/query/domain/Report.java
@@ -5,7 +5,7 @@ import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import pinup.backend.feed.entity.Feed;
+import pinup.backend.feed.command.entity.Feed;
 import pinup.backend.member.command.domain.Admin;
 import pinup.backend.member.command.domain.Users;
 
@@ -19,7 +19,7 @@ public class Report {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "report_id")
-    private Integer reportId;
+    private Long reportId;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "user_id", nullable = false)
